Move JAVA files under 'sygus' package folder
