+++ .bsp/sbt.json
@@ -1,1 +1,1 @@
-{"name":"sbt","version":"1.4.3","bspVersion":"2.0.0-M5","languages":["scala"],"argv":["C:\\Program Files\\Java\\jdk-15.0.1/bin/java","-Xms100m","-Xmx100m","-classpath","C:/Users/shrad/AppData/Roaming/JetBrains/IdeaIC2020.3/plugins/Scala/launcher/sbt-launch.jar","xsbt.boot.Boot","-bsp"]}
+{"name":"sbt","version":"1.6.2","bspVersion":"2.0.0-M5","languages":["scala"],"argv":["C:\\Program Files\\JetBrains\\IntelliJ IDEA 2022.1\\jbr/bin/java","-Xms100m","-Xmx100m","-classpath","C:/Users/boazy/AppData/Roaming/JetBrains/IntelliJIdea2022.1/plugins/Scala/launcher/sbt-launch.jar","xsbt.boot.Boot","-bsp","--sbt-launch-jar=C:/Users/boazy/AppData/Roaming/JetBrains/IntelliJIdea2022.1/plugins/Scala/launcher/sbt-launch.jar"]}
