--- src/main/java/SyGuSLexer.tokens
@@ -1,71 +1,0 @@
-T__0=1
-T__1=2
-T__2=3
-T__3=4
-T__4=5
-T__5=6
-T__6=7
-T__7=8
-T__8=9
-T__9=10
-T__10=11
-T__11=12
-T__12=13
-T__13=14
-T__14=15
-T__15=16
-T__16=17
-T__17=18
-T__18=19
-T__19=20
-T__20=21
-T__21=22
-T__22=23
-T__23=24
-T__24=25
-T__25=26
-T__26=27
-T__27=28
-T__28=29
-T__29=30
-T__30=31
-LineComment=32
-WS=33
-Numeral=34
-Decimal=35
-BoolConst=36
-HexConst=37
-BinConst=38
-StringConst=39
-Symbol=40
-'Int'=1
-'Bool'=2
-'Real'=3
-'('=4
-'BitVec'=5
-')'=6
-'exists'=7
-'forall'=8
-'let'=9
-'grammars'=10
-'fwd-decls'=11
-'recursion'=12
-'check-synth'=13
-'constraint'=14
-'declare-var'=15
-'inv-constraint'=16
-'set-feature'=17
-':'=18
-'synth-fun'=19
-'synth-inv'=20
-'declare-datatype'=21
-'declare-datatypes'=22
-'declare-sort'=23
-'define-fun'=24
-'define-sort'=25
-'set-info'=26
-'set-logic'=27
-'set-option'=28
-'Constant'=29
-'Variable'=30
-'_'=31
