+++ src/main/java/sygus/Python3Lexer.java
@@ -3,11 +3,9 @@ package sygus;
 import org.antlr.v4.runtime.Lexer;
 import org.antlr.v4.runtime.CharStream;
 import org.antlr.v4.runtime.Token;
-import org.antlr.v4.runtime.TokenStream;
 import org.antlr.v4.runtime.*;
 import org.antlr.v4.runtime.atn.*;
 import org.antlr.v4.runtime.dfa.DFA;
-import org.antlr.v4.runtime.misc.*;
 
 @SuppressWarnings({"all", "warnings", "unchecked", "unused", "cast"})
 public class Python3Lexer extends Lexer {
@@ -233,7 +231,7 @@ public class Python3Lexer extends Lexer {
 	}
 
 	@Override
-	public String getGrammarFileName() { return "Python3.g4"; }
+	public String getGrammarFileName() { return "sygus/Python3.g4"; }
 
 	@Override
 	public String[] getRuleNames() { return ruleNames; }
