+++ src/main/java/sygus/ThrowingLexerErrorListener.java
@@ -1,0 +1,27 @@
+package sygus;
+import org.antlr.v4.runtime.BaseErrorListener;
+import org.antlr.v4.runtime.RecognitionException;
+import org.antlr.v4.runtime.Recognizer;
+import org.antlr.v4.runtime.misc.ParseCancellationException;
+
+public class ThrowingLexerErrorListener extends BaseErrorListener
+{
+    @Override
+    public void syntaxError(Recognizer<?, ?> recognizer, Object offendingSymbol, int line, int charPositionInLine, String msg, RecognitionException e) {
+        throw new ParseCancellationException(e);
+    }
+    //    //BaseErrorListener implementation; not called in my test, but left it just in case
+//
+//    public override void SyntaxError(IRecognizer recognizer, IToken offendingSymbol, int line, int charPositionInLine, string msg, RecognitionException e)
+//    {
+//        throw new ArgumentException("Invalid Expression: {0}", msg, e);
+//    }
+//
+//    //IAntlrErrorListener<int> implementation; this one actually gets called.
+//
+//    public void SyntaxError(IRecognizer recognizer, int offendingSymbol, int line, int charPositionInLine, string msg, RecognitionException e)
+//    {
+//        throw new ArgumentException("Invalid Expression: {0}", msg, e);
+//    }
+}
+
